--- a/src/metrics/filters/filter_benchmark_test.go
+++ b/src/metrics/filters/filter_benchmark_test.go
@@ -126,11 +126,13 @@ func BenchmarkRangeFilterRangeMatchNegation(b *testing.B) {
 }
 
 func BenchmarkMultiRangeFilterPrefixCollision(b *testing.B) {
-	benchMultiRangeFilter(b, []byte("test_1,test_1_suffix,extra_unused_pattern"), false, [][]byte{[]byte("test_1"), []byte("test_1_suffix")})
+	benchMultiRangeFilter(b, []byte("test_1,test_1_suffix,extra_unused_pattern"),
+		false, [][]byte{[]byte("test_1"), []byte("test_1_suffix")})
 }
 
 func BenchmarkMultiRangeFilterSuffixCollision(b *testing.B) {
-	benchMultiRangeFilter(b, []byte("test_1,prefix_test_1,extra_unused_pattern"), true, [][]byte{[]byte("test_1"), []byte("prefix_test_1")})
+	benchMultiRangeFilter(b, []byte("test_1,prefix_test_1,extra_unused_pattern"),
+		true, [][]byte{[]byte("test_1"), []byte("prefix_test_1")})
 }
 
 func BenchmarkMultiRangeFilterTwo(b *testing.B) {