--- a/src/metrics/filters/filter_test.go
+++ b/src/metrics/filters/filter_test.go
@@ -350,15 +350,15 @@ func TestMultiCharSequenceFilter(t *testing.T) {
 	validateLookup(t, f, "12test2", true, "12")
 	validateLookup(t, f, "123book", true, "123")
 
-	f, err = newMultiCharSequenceFilter([]byte("arachne_failures,arachne_failures_by_rack"), false)
+	f, err = newMultiCharSequenceFilter([]byte("test,test_post,extra_unused_filter"), false)
 	require.NoError(t, err)
-	validateLookup(t, f, "arachne_failures", true, "")
-	validateLookup(t, f, "arachne_failures_by_rack", true, "")
+	validateLookup(t, f, "test", true, "")
+	validateLookup(t, f, "test_post", true, "")
 
-	f, err = newMultiCharSequenceFilter([]byte("arachne_failures,test_arachne_failures"), true)
+	f, err = newMultiCharSequenceFilter([]byte("test,pre_test,extra_unused_filter"), true)
 	require.NoError(t, err)
-	validateLookup(t, f, "arachne_failures", true, "")
-	validateLookup(t, f, "test_arachne_failures", true, "")
+	validateLookup(t, f, "test", true, "")
+	validateLookup(t, f, "pre_test", true, "")
 }
 
 func validateLookup(t *testing.T, f chainFilter, val string, expectedMatch bool, expectedRemainder string) {