<!--  Thanks for sending a pull request!  Here are some tips for you:
1. If this is your first time, please read our contributor guidelines: https://github.com/m3db/m3/blob/master/CONTRIBUTING.md and developer notes https://github.com/m3db/m3/blob/master/DEVELOPMENT.md
2. Please prefix the name of the pull request with the component you are updating in the format "[component] Change title" (for example "[dbnode] Support out of order writes") and also label this pull request according to what type of issue you are addressing. Furthermore, if this is a WIP or DRAFT PR, please create a draft PR instead: https://github.blog/2019-02-14-introducing-draft-pull-requests/
3. Ensure you have added or ran the appropriate tests for your PR: read more at https://github.com/m3db/m3/blob/master/DEVELOPMENT.md#testing-changes
4. Follow the instructions for writing a changelog note: read more at https://github.com/m3db/m3/blob/master/DEVELOPMENT.md#updating-the-changelog
-->

**What this PR does / why we need it**: Currently, the multiCharSequenceFilter `matches` function returns on the first pattern found [here](https://github.com/m3db/m3/blob/37c5a40d655a7cdde9cbc3725b4f377ded761d53/src/metrics/filters/filter.go#L589-L595) even if there are other patterns available with better matches. This results in a suboptimal byte[] val being returned which causes the public `matches` function to return a false negative [here](https://github.com/m3db/m3/blob/37c5a40d655a7cdde9cbc3725b4f377ded761d53/src/metrics/filters/filter.go#L662). The `len(val)` is not zero because a suboptimal val was returned when a better pattern could have been found if the multiCharSequenceFilter `matches` function iterated through all patterns available and looked for the best match.

<!--
If you have an issue this change addresses, please add the following details:
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
Fixes # https://github.com/m3db/m3/issues/4189

**Special notes for your reviewer**: Please provide any go style improvements or tips you think are relevant

**Does this PR introduce a user-facing and/or backwards incompatible change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".
-->
```release-note
NONE
```

**Does this PR require updating code package or user-facing documentation?**:
<!--
If no, just write "NONE" in the documentation-note block below.
If yes, describe which documentation you updated.
Note: Any changes that significantly updates how a code package is architectured or functions requires code package documentation updates in the form of updates to the README.md file in that package.
-->
```documentation-note
NONE
```
Benchmark Test Results Before and After Changes to `matches` method of `multiCharSequenceFilter`:

| testname                           | old-bench-test-results.txt sec/op | new-bench-test-results.txt sec/op | sec/op vs base         |
|------------------------------------|-----------------------------------|-----------------------------------|------------------------|
| MultiRangeFilterPrefixCollision-10 | 10.61n ± 1%                       | 16.27n ± 4%                       | +53.47% (p=0.000 n=10) |
| MultiRangeFilterSuffixCollision-10 | 11.51n ± 7%                       | 17.00n ± 3%                       | +47.65% (p=0.000 n=10) |
| MultiRangeFilterTwo-10             | 17.48n ± 5%                       | 17.91n ± 2%                       | +2.43% (p=0.004 n=10)  |
| MultiRangeFilterSelectTwo-10       | 21.45n ± 3%                       | 24.30n ± 2%                       | +13.31% (p=0.000 n=10) |
| MultiRangeFilterTrieTwo-10         | 16.44n ± 4%                       | 16.07n ± 2%                       | ~ (p=0.342 n=10)       |
| MultiRangeFilterSix-10             | 38.89n ± 1%                       | 42.51n ± 2%                       | +9.32% (p=0.000 n=10)  |
| MultiRangeFilterSelectSix-10       | 56.47n ± 0%                       | 65.47n ± 1%                       | +15.94% (p=0.000 n=10) |
| MultiRangeFilterTrieSix-10         | 33.67n ± 0%                       | 33.97n ± 4%                       | +0.89% (p=0.005 n=10)  |
| geomean                            | 22.04n                            | 25.57n                            | +16.04%                |