@saad-zaman 2023-02-16
> It's not immediately clear that this is a bug without mentioning how it's used in the public boolean `Matches` function [here](https://github.com/m3db/m3/blob/37c5a40d655a7cdde9cbc3725b4f377ded761d53/src/metrics/filters/filter.go#L662).
> 
> Maybe explicitly mention this in the description and how it would cause false negatives.

Updated in the description and the associated [issue](https://github.com/m3db/m3/issues/4189).