@prateek 2023-02-16
@saad-zaman before landing, can you run existing/make new benchmarks for this change too -- https://dave.cheney.net/2013/06/30/how-to-write-benchmarks-in-go is a good guide to get you going.